% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflow.R
\docType{data}
\name{elmer.example}
\alias{elmer.example}
\title{A DNA methylation matrix and a gene expression matrix for 10 GBM and 10 LGG samples prepared for
the creation of an ELMER object.}
\format{Gene expression: A matrix with 21022 rows and 20 samples, DNA methylation: A matrix with 8066 rows and 20 samples,}
\description{
A DNA methylation matrix and a gene expression matrix for 10 GBM and 10 LGG samples prepared for
the creation of an ELMER object.
}
\keyword{internal}

