% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflow.R
\docType{data}
\name{exp.elmer}
\alias{exp.elmer}
\title{A gene expression matrix for 10 GBM and 10 LGG samples prepared for
the creation of an ELMER object.}
\format{A matrix with 21022 rows and 20 samples}
\description{
A gene expression matrix for 10 GBM and 10 LGG samples prepared for
the creation of an ELMER object.
}
\keyword{internal}

