% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflow.R
\docType{data}
\name{gbm.samples}
\alias{gbm.samples}
\title{Identifiers for the 10 GBM samples in the ELMER objects}
\format{A vector of 10 barcodes}
\description{
Identifiers for the 10 GBM samples in the ELMER objects
}
\keyword{internal}

