% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflow.R
\docType{data}
\name{GBMnocnvhg19}
\alias{GBMnocnvhg19}
\title{CNV data for 20 TCGA-GBM samples}
\format{A matrix with 3252 rows and 6 columns}
\description{
CNV data for 20 TCGA-GBM samples
}
\keyword{internal}

