% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.adjacency.biogrid}
\alias{get.adjacency.biogrid}
\title{Get a matrix of interactions of genes from biogrid}
\usage{
get.adjacency.biogrid(tmp.biogrid, names.genes = NULL)
}
\arguments{
\item{tmp.biogrid}{Biogrid table}

\item{names.genes}{List of genes to filter from output. Default: consider all genes}
}
\value{
A matrix with 1 for genes that interacts, 0 for no interaction.
}
\description{
Using biogrid database, it will create a matrix of gene interations.
If columns A and row B has value 1, it means the gene A and gene B interatcs.
}

