% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflow.R
\docType{data}
\name{LGGIlumina_Hiseq}
\alias{LGGIlumina_Hiseq}
\title{A gene expression matrix wih 20 LGG samples}
\format{Gene expression: A SummariedExperiment object with 21022 rows and 20 columns}
\description{
A gene expression matrix wih 20 LGG samples
}
\keyword{internal}

