% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matched_met_exp}
\alias{matched_met_exp}
\title{Get GDC samples with both DNA methylation and Gene expression data}
\usage{
matched_met_exp("TCGA-ACC")
}
\arguments{
\item{project}{A GDC project}

\item{n}{Number of samples to return. If NULL return all (default)}
}
\value{
A vector of barcodes
}
\description{
For a given TCGA project it gets the  samples (barcode) with both DNA methylation and Gene expression data
}
\examples{
# Get ACC samples with both DNA methylation and gene expression
samples <- matched_met_exp("TCGA-ACC")
}

