% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflow.R
\docType{data}
\name{met.elmer}
\alias{met.elmer}
\title{A DNA methylation matrix for 10 GBM and 10 LGG samples prepared for
the creation of an ELMER object.}
\format{A matrix with 8066 rows and 20 samples}
\description{
A DNA methylation matrix for 10 GBM and 10 LGG samples prepared for
the creation of an ELMER object.
}
\keyword{internal}

