% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TCGAsurvival_TFplot}
\alias{TCGAsurvival_TFplot}
\title{Evaluate survival based on the levels of TF expression}
\usage{
TCGAsurvival_TFplot(TF, mee, clinical, percentage = 0.3)
}
\arguments{
\item{TF}{A transcription factor}

\item{mee}{A mee object from ELMER}

\item{clinical}{Clinical data from GDC}

\item{percentage}{Percentage of samples of each group. Default 30 percent in the higher expression 
group and  30 percent in the lower expression group}
}
\value{
A pdf with the survival plot
}
\description{
This function is an auxiliar function used to integrate ELMER and TCGAbiolinks pacakges.
For a given regulatory TF found by ELMER, we will use TCGAbiolinks to get the clinical data of the
samples and evaluate the survival for two groups, ones with higher expression of that TF and ones
with lower expression.
}

